% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_hazard_matrix}
\alias{assert_hazard_matrix}
\title{Assert hazard matrix}
\usage{
assert_hazard_matrix(x)
}
\arguments{
\item{x}{(\code{matrix()})\cr
A matrix of hazards.
Rows are observations and columns are (increasing) time points.}
}
\value{
if the assertion fails an error occurs, otherwise \code{NULL} is returned
invisibly.
}
\description{
Asserts if the given input is a (discrete) hazard matrix.
The following checks are performed:
\enumerate{
\item All values are non-negative, i.e. \eqn{h(t) \ge 0}
\item Column names correspond to time points and should therefore be coercable to
\code{numeric} and increasing
}
}
\examples{
x = matrix(data = c(0,0.2,0.1,0.5,0.05,0.7), nrow = 2, ncol = 3, byrow = TRUE)
colnames(x) = c(12, 34, 42)
x

assert_hazard_matrix(x)

}
