% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDistr.R
\name{cdf}
\alias{cdf}
\alias{cdf.survDistr}
\title{Cumulative Distribution Function}
\usage{
cdf(x, ...)

\method{cdf}{survDistr}(x, times = NULL)
}
\arguments{
\item{x}{(\code{survDistr|...})\cr
An object used to select the appropriate method.}

\item{times}{(\code{numeric()})\cr
An optional numeric vector of time points.
Time points don't have to be sorted or unique, just non-negative numbers.}
}
\value{
A \code{matrix} of CDF values.
}
\description{
Computes the cumulative distribution function \eqn{F(t)} at the specified time points.
\eqn{F(t)} is the probability that the event has occurred up until time \eqn{t}.
}
\details{
If \code{x} is a \link{survDistr}, and no time points are specified, returns
\eqn{F(t) = 1 - S(t)}, where \eqn{S(t)} is the stored survival matrix.
}
\seealso{
Other methods: 
\code{\link{cumhazard}()},
\code{\link{hazard}()},
\code{\link{pdf}()},
\code{\link{survival}()},
\code{\link{times}()}
}
\concept{methods}
