% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDistr.R
\name{hazard}
\alias{hazard}
\alias{hazard.survDistr}
\title{Hazard function}
\usage{
hazard(x, ...)

\method{hazard}{survDistr}(x, times = NULL, eps = 1e-06)
}
\arguments{
\item{x}{(\code{survDistr|...})\cr
An object used to select the appropriate method.}

\item{times}{(\code{numeric()})\cr
An optional numeric vector of time points.
Time points don't have to be sorted or unique, just non-negative numbers.}

\item{eps}{(\code{numeric(1)})\cr
Very small number to substitute zero values in order to prevent errors
in e.g. log(0) and/or division-by-zero calculations.
Default value is 1e-06.}
}
\value{
A \code{matrix} of hazards.
}
\description{
Computes the hazard at the specified time points as: \eqn{h(t) = H(t) - H(t-1)}.
}
\details{
If \code{x} is a \link{survDistr}, and no time points are specified, the calculation of
\eqn{H(t)} uses the stored survival \code{matrix} \eqn{S(t)}.
}
\seealso{
Other methods: 
\code{\link{cdf}()},
\code{\link{cumhazard}()},
\code{\link{pdf}()},
\code{\link{survival}()},
\code{\link{times}()}
}
\concept{methods}
