% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_funs.R
\name{interp_cif}
\alias{interp_cif}
\title{Interpolate CIF matrix}
\usage{
interp_cif(x, times = NULL, eval_times = NULL, add_times = TRUE, check = TRUE)
}
\arguments{
\item{x}{(\code{matrix()})\cr
CIF matrix (rows = observations, columns = time points).}

\item{times}{(\code{numeric()} | \code{NULL})\cr
Anchor time points. If \code{NULL}, extracted from \code{colnames(x)}.}

\item{eval_times}{(\code{numeric()} | \code{NULL})\cr
New time points at which to interpolate.
Values need to be unique, increasing non-negative numbers.
If \code{NULL}, \code{x} is returned unchanged.}

\item{add_times}{(\code{logical(1)})\cr
If \code{TRUE}, attach \code{eval_times} as colnames in the output matrix.}

\item{check}{(\code{logical(1)})\cr
If \code{TRUE}, run input matrix validation checks using \code{\link[=assert_prob]{assert_prob()}};
set to \code{FALSE} to skip checks (NOT recommended for external use).}
}
\value{
Interpolated CIF matrix.
}
\description{
Interpolates cumulative incidence (CIF) functions (corresponding to one competing event only)
using left-continuous constant interpolation.
}
