% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumgg.R
\name{KumGG}
\alias{KumGG}
\alias{dkumgg}
\alias{pkumgg}
\alias{qkumgg}
\alias{hkumgg}
\alias{Hkumgg}
\alias{rkumgg}
\alias{ml.kumgg}
\title{Kumaraswamy Generalized Gamma Distribution}
\source{
The source code of all distributions in this package can also be
found on the \href{https://github.com/aneisse/survdistr}{survdistr} Github repository.
}
\usage{
dkumgg(x, tau, alpha = 1, k, lambda, phi, log = FALSE)

pkumgg(q, tau, alpha = 1, k, lambda, phi, lower.tail = TRUE,
  log.p = FALSE)

qkumgg(p, tau, alpha = 1, k, lambda, phi, lower.tail = TRUE,
  log.p = FALSE)

hkumgg(q, tau, alpha = 1, k, lambda, phi)

Hkumgg(q, tau, alpha = 1, k, lambda, phi)

rkumgg(n, tau, alpha = 1, k, lambda, phi, cens.prop = 0)

ml.kumgg(x, tau.ini, alpha.ini, k.ini, lambda.ini, phi.ini)
}
\arguments{
\item{x, q}{numeric vector of quantiles.}

\item{alpha}{scale parameter \eqn{\alpha > 0}.}

\item{lambda}{shape parameter \eqn{\lambda > 0}.}

\item{phi}{shape parameter \eqn{\phi \ge 0}.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities/densities \code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE}, probabilities are \eqn{P[X \le x]}, otherwise, \eqn{P[X \ge x]}}

\item{n}{desired size of the random number sample.}

\item{cens.prop}{proportion of censored data to be simulated. If greater than \code{0}, a matrix will be
returned instead of a vector. The matrix will contain the random values and a censorship indicator variable.}

\item{gamma}{shape parameter \eqn{\gamma > 0}.}
}
\value{
\code{dkumgg} gives the density, \code{pkumgg} gives the distribution
function, \code{qkumgg} gives the quantile function, and \code{rkumgg} generates random values.

The length of the result is determined by \code{n} for \code{rkumgg}, for the other fucntions the
length is the same as the vector passed to the first argument.

Only the first element of the logical arguments are used.
}
\description{
Density, distribution function, quantile function and random generation
for the KumGG distribution with parameters \code{lambda}, \code{phi},
\code{tau}, \code{alpha} and \code{k}.
}
\details{
The KumGG distribution was described by Ortega \emph{et al} (2011) and has density

\deqn{f(x) = (\lambda\phi\tau)/(\alpha\Gamma[k])(x/\alpha)^(\tauk-1)e^(-(x/\alpha)^\tau)
(\gamma[k, (x/\alpha)^\tau])^(\lambda-1)(1-(\gamma[k, (x/\alpha)^\tau])^\lambda)^(\phi-1)}

where \eqn{\gamma[., .]} is the incomplete gamma ratio and \eqn{\Gamma[.]} is the gamma funcion.
The scale parameter is \eqn{\alpha}, the shape parameters are \eqn{\lambda}, \eqn{\phi} and
\eqn{\tau} and \eqn{k}. The parameters \eqn{\lambda} and \eqn{phi}, come from the Kumaraswamy
Generalized family introduced by Cordeiro and Castro (2011).

With \code{phi = 1} KumGG becomes the Exponentiated Generalized Gamma distribution described
by Cordeiro \emph{et al} (2011). Additionally, if \code{tau = k = 1} the Exponentiated Exponential.

When \code{k = 1} then the KumGG distribution becomes the KumW distribution
described by Cordeiro \emph{et al} (2010). For \code{k = 1} the KumGG becomes KumG distribution.

The above are arguably the most important sub-models of KumGG. More sub-models are described in
Ortega \emph{et al} (2011).
}
\examples{
# Generating values and comparing with the function
x <- rkumgg(10000, tau = 0.9, alpha = 2, k = 0.5, lambda = 3, phi = 10)
hist(x, probability = T, breaks = 100)
curve(dkumgg(x, tau = 0.9, alpha = 2, k = 0.5, lambda = 3, phi = 10),
      from = 0, to = 3, add = T)

}
\references{
ORTEGA, E. M. M.; CORDEIRO, G. M.; PASCOA, M. A. R. The generalized gamma
geometric distribution. Journal of Statistical Theory and Applications,
2011, 10.3: 433-454.

CORDEIRO, G. M.; ORTEGA, E. M. M; SILVA, G. O. The exponentiated generalized
gamma distribution with application to lifetime data. Journal of statistical
computation and simulation, 2011, 81.7: 827-842.

CORDEIRO, G. M.; ORTEGA, E. M. M; NADARAJAH, S.. The Kumaraswamy Weibull
distribution with application to failure data. Journal of the Franklin
Institute, 2010, 347.8: 1399-1429.

CORDEIRO, G. M.; DE CASTRO, M. A new family of generalized distributions. Journal of
statistical computation and simulation, 2011, 81.7: 883-898.
}
\seealso{
LINK TO OTHER PACKAGE DISTRIBUTIONS
}
\author{
Anderson Neisse <a.neisse@gmail.com>
}
