% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumll.R
\name{KumLL}
\alias{KumLL}
\alias{dkumll}
\alias{pkumll}
\alias{qkumll}
\alias{hkumll}
\alias{rkumll}
\title{Kumarasuammy Log-Logistic Distribution}
\source{
The source code of all distributions in this package can also be
found on the \href{https://github.com/aneisse/survdistr}{survdistr} Github repository.
}
\usage{
dkumll(x, alpha, gamma, lambda, phi, log = FALSE)

pkumll(q, alpha, gamma, lambda, phi, lower.tail = TRUE, log.p = FALSE)

qkumll(p, alpha, gamma, lambda, phi, lower.tail = TRUE, log.p = FALSE)

hkumll(q, alpha, gamma, lambda, phi)

rkumll(n, alpha, gamma, lambda, phi, cens.prop = 0)
}
\arguments{
\item{x, q}{numeric vector of quantiles.}

\item{alpha}{scale parameter \eqn{\alpha > 0}.}

\item{gamma}{shape parameter \eqn{\gamma > 0}.}

\item{lambda}{shape parameter \eqn{\lambda > 0}.}

\item{phi}{shape parameter \eqn{\phi \ge 0}.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities/densities \code{p} are given as \code{log(p)}.}

\item{lower.tail}{logical; if \code{TRUE}, probabilities are \eqn{P[X \le x]}, otherwise, \eqn{P[X \ge x]}}

\item{n}{desired size of the random number sample.}

\item{cens.prop}{proportion of censored data to be simulated. If greater than \code{0}, a matrix will be
returned instead of a vector. The matrix will contain the random values and a censorship indicator variable.}
}
\value{
\code{dkumll} gives the density, \code{pkumll} gives the distribution
function, \code{qkumll} gives the quantile function, and \code{rkumll} generates random values.

The length of the result is determined by \code{n} for \code{rkumll}, for the other fucntions the
length is the same as the vector passed to the first argument.

Only the first element of the logical arguments are used.
}
\description{
Density, distribution function, quantile function and random generation
for the KumLL distribution with parameters \code{lambda}, \code{phi},
\code{c}, \code{k} and \code{s}.
}
\details{
The KumLL distribution was described by Santana \emph{et al} (2012) and has density

\deqn{f(x) = (\lambda\phi\gamma)/(\alpha^(\lambda\gamma))x^(\lambda\gamma-1)
(1+(t/\alpha)^\gamma)^(-\lambda-1)(1-(1-1/(1+(t/\alpha)^\gamma))^\lambda)^(\phi-1)}
with scale parameter \eqn{\alpha}, shape parameters \eqn{\lambda}, \eqn{\phi} and
\eqn{\gamma} that govern the distribution's skewness. The parameters \eqn{\lambda}
and \eqn{phi}, come from the Kumaraswamy Generalized family introduced by Cordeiro
and Castro (2011).

The KumLL is a special case of KumBII introduced by Parnaíba \emph{et al} (2013).

With \code{phi = 1} KumLL becomes the Exponentiated Log-Logistic distribution.
In addition, when \code{lambda = 1} it becomes the Log-Logistic distribution.
Those are arguably the most important sub-models to KumLL.

When \code{lambda = 1} then the KumLL distribution becomes the BXII distribution
described by Zimmer \emph{et al} (1998).

This distribution's failure rate function accommodates increasing, decreasing, unimodal
and bathtub shaped forms, that depend basically on the values of the shape parameters.
Moreover, it is quite flexible for modeling survival data.
}
\examples{
# Generating values and comparing with the function
x <- rkumll(10000, alpha = 0.5, gamma = 2, lambda = 2, phi = 2)
hist(x, probability = T, breaks = 100)
curve(dkumll(x, alpha = 0.5, gamma = 2, lambda = 2, phi = 2),
      from = 0, to = 25, add = T)

}
\references{
DE SANTANA, T. V. F.; Ortega, E. M.; Cordeiro, G. M.;  Silva, G. O. The
Kumaraswamy-log-logistic distribution. Journal of Statistical Theory
and Applications, 2012, 11.3: 265-291.

PARANAÍBA, P. F.; Ortega, E. M.; Cordeiro, G. M.; Pascoa, M. A. D. The Kumaraswamy
Burr XII distribution: theory and practice. Journal of Statistical Computation
and Simulation, 2013, 83.11: 2117-2143.

CORDEIRO, G. M.; DE CASTRO, M. A new family of generalized distributions. Journal of
statistical computation and simulation, 2011, 81.7: 883-898.

ZIMMER, W. J.; KEATS, J. B.; WANG, F. K. The Burr XII distribution in reliability
analysis. Journal of quality technology, 1998, 30.4: 386-394.
}
\seealso{
LINK TO OTHER PACKAGE DISTRIBUTIONS
}
\author{
Anderson Neisse <a.neisse@gmail.com>
}
