% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDistr.R
\name{pdf}
\alias{pdf}
\alias{pdf.survDistr}
\title{Probability Density Function}
\usage{
pdf(x, ...)

\method{pdf}{survDistr}(x, times = NULL)
}
\arguments{
\item{x}{(\code{survDistr|...})\cr
An object used to select the appropriate method.}

\item{times}{(\code{numeric()})\cr
An optional numeric vector of time points.
Time points don't have to be sorted or unique, just non-negative numbers.}
}
\value{
A \code{matrix} of PDF values.
}
\description{
Computes the probability density function \eqn{f(t)} at the specified time points.
\eqn{f(t)} is the probability of the event occurring at the specific time \eqn{t}.
}
\details{
If \code{x} is a \link{survDistr}, and no time points are specified, returns
\eqn{f(t) = F(t) - F(t-1)}, where \eqn{F(t)} is the cumulative distribution
function calculated directly from the stored survival matrix.
If new time points are specified, these are first ordered and duplicated values
are removed, so that the \eqn{F(t_{new})} can be calculated (always using linear
interpolation to avoid zero values for the pdf for time points lying in between
the ones we have S(t) in the stored survival matrix). pdf values for t_new < min(times)
or t_new > max(times) will be zero nonetheless.
}
\seealso{
Other methods: 
\code{\link{cdf}()},
\code{\link{cumhazard}()},
\code{\link{hazard}()},
\code{\link{survival}()},
\code{\link{times}()}
}
\concept{methods}
