% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDistr.R
\name{print.survDistr}
\alias{print.survDistr}
\title{Print method for survDistr}
\usage{
\method{print}{survDistr}(x, ...)
}
\arguments{
\item{x}{(\code{survDistr|...})\cr
An object used to select the appropriate method.}
}
\description{
Displays summary information about a \link{survDistr} object, including
the number of observations and time points.
}
