% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDistr.R
\name{survival}
\alias{survival}
\alias{survival.survDistr}
\title{Survival function}
\usage{
survival(x, ...)

\method{survival}{survDistr}(x, times = NULL)
}
\arguments{
\item{x}{(\code{survDistr|...})\cr
An object used to select the appropriate method.}

\item{times}{(\code{numeric()})\cr
An optional numeric vector of time points.
Time points don't have to be sorted or unique, just non-negative numbers.}
}
\value{
A \code{matrix} of survival probabilities.
}
\description{
Computes survival probabilities \eqn{S(t)} at the specified time points.
}
\details{
If \code{x} is a \link{survDistr}, and no time points are specified, returns the
stored survival \code{matrix}.
}
\seealso{
Other methods: 
\code{\link{cdf}()},
\code{\link{cumhazard}()},
\code{\link{hazard}()},
\code{\link{pdf}()},
\code{\link{times}()}
}
\concept{methods}
