% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDistr.R
\name{times}
\alias{times}
\alias{times.survDistr}
\title{Time Points}
\usage{
times(x, ...)

\method{times}{survDistr}(x, ...)
}
\arguments{
\item{x}{(\code{survDistr|...})\cr
An object used to select the appropriate method.}
}
\value{
A \code{vector} of numeric values.
}
\description{
Returns the time points associated with the object \code{x}.
}
\seealso{
Other methods: 
\code{\link{cdf}()},
\code{\link{cumhazard}()},
\code{\link{hazard}()},
\code{\link{pdf}()},
\code{\link{survival}()}
}
\concept{methods}
