% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_survDistr.R
\name{as_survDistr}
\alias{as_survDistr}
\alias{as_survDistr.matrix}
\alias{as_survDistr.survDistr}
\alias{as_survDistr.default}
\title{Coerce Object to \link{survDistr}}
\usage{
as_survDistr(x, ...)

\method{as_survDistr}{matrix}(
  x,
  times = NULL,
  method = "const_surv",
  check = TRUE,
  trim_duplicates = FALSE,
  ...
)

\method{as_survDistr}{survDistr}(x, ...)

\method{as_survDistr}{default}(x, ...)
}
\arguments{
\item{x}{Object to convert.}

\item{...}{Additional arguments passed to methods.}

\item{times}{(\code{numeric})
Numeric vector of time points corresponding to columns of \code{x}.
If \code{NULL}, column names of \code{x} are used.}

\item{method}{(\code{character(1)})\cr
Interpolation method passed to \link{survDistr} constructor.}

\item{check}{(\code{logical(1)})\cr
Whether to validate \code{x} and \code{times}.}

\item{trim_duplicates}{(\code{logical(1)})\cr
Whether to remove duplicate S(t) values and corresponding time points.}
}
\value{
A \link{survDistr} object.
}
\description{
S3 generic to coerce supported objects into a \link{survDistr} object.
}
