% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_surv.R
\name{convert_to_surv}
\alias{convert_to_surv}
\title{Convert density/hazard to survival}
\usage{
convert_to_surv(
  x,
  times = NULL,
  input = "cont_haz",
  check = TRUE,
  clamp_surv = FALSE,
  eps = 1e-06
)
}
\arguments{
\item{x}{(\code{numeric()} | \code{matrix()})\cr
Input vector or matrix (rows = observations, columns = time points).}

\item{times}{(\code{numeric()} | \code{NULL})\cr
Anchor time points. If \code{NULL}, extracted from names/colnames of \code{x}.}

\item{input}{(\code{character(1)})\cr
Input type. One of \code{"disc_haz"}, \code{"disc_dens"}, \code{"cont_haz"} or \code{"cont_dens"}.}

\item{check}{(\code{logical(1)})\cr
If \code{TRUE} (default), run \emph{input} validation checks.
Disable only if you know the input is valid and want to skip checks for speed.}

\item{clamp_surv}{(\code{logical(1)})\cr
If \code{TRUE}, clamp survival probabilities to \verb{[eps, 1]} to avoid numerical issues.}

\item{eps}{(\code{numeric(1)})\cr
Small value used to clamp near-zero survival probabilities if \code{clamp_surv = TRUE}.}
}
\value{
A numeric vector or matrix of survival probabilities with the same
dimensions as \code{x}.
}
\description{
Converts density or hazards from one of four input representations to survival
probabilities at the same anchor time points (no interpolation).
}
\details{
Let \eqn{t_1,\dots,t_B} denote the anchor time points,
\eqn{\Delta_j = t_j - t_{j-1}}, and \eqn{S_j = S(t_j)} the survival
probabilities at the anchors. The conversion depends on the value
of \code{input}:
\itemize{
\item \code{"disc_dens"}: \eqn{S_j = 1 - \sum_{k=1}^j \tilde f_k}
\item \code{"disc_haz"}: \eqn{S_j = \prod_{k=1}^j (1 - \tilde h_k)}
\item \code{"cont_dens"}: \eqn{S_j = 1 - \sum_{k=1}^j f_k \Delta_k}
\item \code{"cont_haz"}: \eqn{S_j = \exp\!\left(-\sum_{k=1}^j \lambda_k \Delta_k\right)}
}
}
\section{Validation}{

If \code{check = TRUE}, we validate that the input is a proper discrete density/hazard matrix
or vector using \code{\link[=assert_prob]{assert_prob()}}.
For continuous hazards/densities, we only check that the input is a non-negative numeric
matrix/vector.
}

\examples{
# Continuous hazard => survival
haz_cont = c(0.02, 0.1, 0.2, 0.15)
times = c(0, 1, 2, 3)
convert_to_surv(haz_cont, times = times, input = "cont_haz")

# Discrete hazard => survival
haz_disc = c(0.1, 0.2, 0.15)
times = c(1, 2, 3)
convert_to_surv(haz_disc, times = times, input = "disc_haz")

}
