% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp.R
\name{interp}
\alias{interp}
\title{Interpolate Survival Curves}
\usage{
interp(
  x,
  times = NULL,
  eval_times = NULL,
  method = "const_surv",
  output = "surv",
  add_times = TRUE,
  check = TRUE,
  eps = 1e-12,
  trim_duplicates = FALSE
)
}
\arguments{
\item{x}{(\code{numeric()} | \code{matrix()})\cr
Survival vector or matrix (rows = observations, columns = time points).}

\item{times}{(\code{numeric()} | \code{NULL})\cr
Anchor time points. If \code{NULL}, extracted from names/colnames.}

\item{eval_times}{(\code{numeric()} | \code{NULL})\cr
New time points at which to interpolate.
Values need to be unique, increasing non-negative numbers.
If \code{NULL}, the anchor \code{times} are used.}

\item{method}{(\code{character(1)})\cr
Interpolation method; one of \code{"const_surv"} (default), \code{"const_dens"} (alias: \code{"linear_surv"})
and \code{"const_haz"} (alias: \code{"exp_surv"}).}

\item{output}{(\code{character(1)})\cr
Output type: \code{"surv"}, \code{"cdf"}, \code{"cumhaz"}, \code{"density"} or \code{"hazard"}.}

\item{add_times}{(\code{logical(1)})\cr
If \code{TRUE}, attach \code{eval_times} as names/colnames.}

\item{check}{(\code{logical(1)})\cr
If \code{TRUE}, run input matrix validation checks using \code{\link[=assert_prob]{assert_prob()}};
set to \code{FALSE} to skip checks (NOT recommended for external use).}

\item{eps}{(\code{numeric(1)})\cr
Small positive value used to replace extremely low survival probabilities when computing cumulative hazard,
preventing numerical instability in \eqn{-\log S(t)} calculations.}

\item{trim_duplicates}{(\code{logical(1)})\cr
If \code{TRUE}, removes adjacent duplicate values from the input using \code{\link[=trim_duplicates]{trim_duplicates()}}.
This eliminates flat segments in survival curves and improves interpolation efficiency.
Default is \code{FALSE}.}
}
\value{
A numeric vector or matrix of interpolated values.
}
\description{
Interpolates survival curves (vector or matrix) at new time points using
internal C++ interpolation functions.
Input must always be \emph{survival probabilities}.
Output can be the survival, cumulative distribution, or density functions, as well as
the hazard or cumulative hazard functions.
}
\examples{
x = matrix(c(1, 0.8, 0.6,
             1, 0.7, 0.4),
           nrow = 2, byrow = TRUE)
times = c(0, 10, 20)
eval_times = c(5, 15, 25)

# constant S(t) interpolation
interp(x, times, eval_times)
# linear S(t) interpolation
interp(x, times, eval_times, method = "linear_surv")
# exponential S(t) interpolation (same as `method = "const_haz"`)
interp(x, times, eval_times, method = "exp_surv")

# cumulative distribution with linear S(t) interpolation
interp(x, times, eval_times, method = "linear_surv", output = "cdf")

# H(t) with linear interpolation
interp(x, times, eval_times, method = "linear_surv", output = "cumhaz")

}
