% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDistr.R
\name{survDistr}
\alias{survDistr}
\title{Constructor for survDistr}
\usage{
survDistr(x, data_type = "surv", inter_type = "const_surv", ...)
}
\arguments{
\item{x}{(\code{matrix()})\cr
A numeric matrix of either survival probabilities (values between 0 and 1) or
hazard values (non-negative values).
Column names must correspond to time points.}

\item{data_type}{(\code{character(1)})\cr
The type of input data.
Either a survival matrix (\code{"survival"}, default) or a hazard (\code{"hazard"}) matrix.}

\item{inter_type}{(\code{character(1)})\cr
The type of interpolation to use for time points different than the ones in
the stored object (accessible via the \code{\link[=times]{times()}} method).
Currently supported interpolation methods include \code{"const_surv"} (default),
\code{"linear_surv"} and \code{"const_haz"}. See details.}

\item{...}{currently not used}
}
\value{
An object of class \link{survDistr} containing the survival matrix.
}
\description{
Creates a \link{survDistr} object, a specialized container designed for storing
and managing prediction outputs from survival models in single-event settings.
This includes models such as Cox proportional hazards, random survival forests,
and other classical or machine learning-based survival estimators.

The main prediction data type can be a survival or a hazard matrix, where
\strong{rows represent observations and columns represent time points}.
}
\details{
The input matrix (survival probabilities \eqn{S(t)} or hazard \eqn{h(t)})
is stored in the \verb{$data} slot, while the interpolation type for subsequent S3
methods is stored in the \verb{$inter_type} slot.

During construction, the functions \code{\link[=assert_surv_matrix]{assert_surv_matrix()}} or \code{\link[=assert_hazard_matrix]{assert_hazard_matrix()}}
are used to validate the input data matrix according to the given \code{data_type}.
}
\section{Available S3 Methods}{

\itemize{
\item \code{\link[=print]{print()}}
\item \code{\link[=times]{times()}}
\item \code{\link[=survival]{survival()}}
\item \code{\link[=cumhazard]{cumhazard()}}
\item \code{\link[=hazard]{hazard()}}
\item \code{\link[=cdf]{cdf()}}
\item \code{\link[=pdf]{pdf()}}
}
}

\examples{
x = matrix(data = c(1,0.6,0.4,0.8,0.8,0.7), nrow = 2,
           ncol = 3, byrow = TRUE)
colnames(x) = c(12, 34, 42)
x = survDistr(x)
x

# stored survival matrix
x$data

# type of interpolation to use in S3 methods
x$inter_type

# time points
times(x)

# S(t) at given time points (constant interpolation)
survival(x, times = c(10, 30, 42, 45))
# same but with linear interpolation
x$inter_type = "linear_surv"
survival(x, times = c(10, 30, 42, 45))
# time points can be unordered and duplicated
survival(x, times = c(10,30,10,50))

# Cumulative hazard
cumhazard(x)

# hazard
hazard(x)

# cumulative distribution function
cdf(x)

# probability density function
pdf(x)

}
