% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDistr.R
\name{survDistr}
\alias{survDistr}
\title{Survival Distribution Container}
\description{
\link{survDistr} is an \link[R6:R6Class]{R6} specialized container designed for storing
and managing prediction outputs from survival models in single-event settings.
This includes models such as Cox proportional hazards, random survival forests,
and other classical or machine learning-based survival estimators.

The main prediction data type can be a survival or a hazard matrix, where
\strong{rows represent observations and columns represent time points}.
}
\details{
The input matrix (survival probabilities \eqn{S(t)} or hazard \eqn{h(t)})
is stored internally and accessed by the \verb{$data} field.
the interpolation type needed for the
public methods is stored in the \verb{$interp_meth} slot.

During construction, the function \code{\link[=assert_prop_matrix]{assert_prop_matrix()}} is used to validate
the input data matrix according to the given \code{data_type}.
}
\examples{
# generate survival matrix
mat = matrix(data = c(1,0.6,0.4,0.8,0.8,0.7), nrow = 2,
             ncol = 3, byrow = TRUE)
times = c(12, 34, 42)
x = survDistr$new(mat, times)
x

# stored survival matrix
x$data()

# interpolation method
x$interp_meth

# time points
x$times

# S(t) at given time points (constant interpolation)
x$survival(times = c(10, 30, 42, 50))
# same but with linear interpolation
x$interp_meth = "linear_surv"
x$survival(times = c(10, 30, 42, 50))
# time points can be unordered and duplicated
x$survival(times = c(10, 30, 10, 50))

# Cumulative hazard
x$cumhazard()

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{times}}{(\code{numeric}])\cr
Numeric vector of time points corresponding to columns of \code{data}.}

\item{\code{data_type}}{(\code{character(1)})\cr
Either \code{"surv"} for survival or \code{"haz"} for hazard matrices.}

\item{\code{interp_meth}}{(\code{character(1)})\cr
Interpolation method; one of \code{"const_surv"}, \code{"linear_surv"}, or \code{"const_haz"}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-survDistr-new}{\code{survDistr$new()}}
\item \href{#method-survDistr-print}{\code{survDistr$print()}}
\item \href{#method-survDistr-data}{\code{survDistr$data()}}
\item \href{#method-survDistr-survival}{\code{survDistr$survival()}}
\item \href{#method-survDistr-cdf}{\code{survDistr$cdf()}}
\item \href{#method-survDistr-cumhazard}{\code{survDistr$cumhazard()}}
\item \href{#method-survDistr-hazard}{\code{survDistr$hazard()}}
\item \href{#method-survDistr-pdf}{\code{survDistr$pdf()}}
\item \href{#method-survDistr-clone}{\code{survDistr$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-survDistr-new"></a>}}
\if{latex}{\out{\hypertarget{method-survDistr-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survDistr$new(
  x,
  times = NULL,
  data_type = "surv",
  interp_meth = "const_surv",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{matrix})\cr
A numeric matrix of either survival probabilities (values between 0 and 1) or
hazard values (non-negative values).
Column names must correspond to time points if \code{times} is \code{NULL}.}

\item{\code{times}}{(\code{numeric(1)})\cr Numeric vector of time points for matrix \code{x},
must match the number of columns.}

\item{\code{data_type}}{(\code{character(1)})\cr
The type of input data.
Either a survival matrix (\code{"survival"}, default) or a hazard (\code{"hazard"}) matrix.}

\item{\code{interp_meth}}{(\code{character(1)})\cr
Interpolation method to use when requesting the quantity of interest at time
points different than the ones in the stored object (accessible via the \code{times} method).
Currently supported interpolation methods include \code{"const_surv"} (default),
\code{"linear_surv"} and \code{"const_haz"}. See details.}

\item{\code{...}}{currently not used}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-survDistr-print"></a>}}
\if{latex}{\out{\hypertarget{method-survDistr-print}{}}}
\subsection{Method \code{print()}}{
Displays summary information about a \link{survDistr} object, including
the number of observations and time points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survDistr$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-survDistr-data"></a>}}
\if{latex}{\out{\hypertarget{method-survDistr-data}{}}}
\subsection{Method \code{data()}}{
Return the stored data matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survDistr$data(add_times = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add_times}}{(\code{logical(1)})\cr
If \code{TRUE}, column names are set to the relevant time points.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{matrix})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-survDistr-survival"></a>}}
\if{latex}{\out{\hypertarget{method-survDistr-survival}{}}}
\subsection{Method \code{survival()}}{
Computes survival probabilities \eqn{S(t)} at the specified time points.
Uses \code{\link[=mat_interp]{mat_interp()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survDistr$survival(times = NULL, add_times = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{times}}{(\code{numeric})\cr
New time points at which to interpolate.
Values do not need to be sorted or unique, just non-negative.
If \code{NULL}, the object's stored time points are used.}

\item{\code{add_times}}{(\code{logical(1)})\cr
If \code{TRUE}, column names are set to the relevant time points.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{matrix} of survival probabilities
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-survDistr-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-survDistr-cdf}{}}}
\subsection{Method \code{cdf()}}{
Computes the cumulative distribution function \eqn{F(t) = 1 - S(t)} at the specified time points.
\eqn{F(t)} is the probability that the event has occurred up until time \eqn{t}.
Uses \code{\link[=mat_interp]{mat_interp()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survDistr$cdf(times = NULL, add_times = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{times}}{(\code{numeric})\cr
New time points at which to interpolate.
Values do not need to be sorted or unique, just non-negative.
If \code{NULL}, the object's stored time points are used.}

\item{\code{add_times}}{(\code{logical(1)})\cr
If \code{TRUE}, column names are set to the relevant time points.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a cdf \code{matrix}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-survDistr-cumhazard"></a>}}
\if{latex}{\out{\hypertarget{method-survDistr-cumhazard}{}}}
\subsection{Method \code{cumhazard()}}{
Computes the cumulative hazard at the specified time points as:
\eqn{H(t) = -log(S(t))}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survDistr$cumhazard(times = NULL, add_times = TRUE, eps = 1e-06)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{times}}{(\code{numeric})\cr
New time points at which to interpolate.
Values do not need to be sorted or unique, just non-negative.
If \code{NULL}, the object's stored time points are used.}

\item{\code{add_times}}{(\code{logical(1)})\cr
If \code{TRUE}, column names are set to the relevant time points.}

\item{\code{eps}}{(\code{numeric(1)})\cr
Very small number to substitute zero values in order to prevent errors
in e.g. log(0) and/or division-by-zero calculations.
Default value is 1e-06.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{matrix} of cumulative hazards.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-survDistr-hazard"></a>}}
\if{latex}{\out{\hypertarget{method-survDistr-hazard}{}}}
\subsection{Method \code{hazard()}}{
Computes the hazard at the specified time points as: \eqn{h(t) = H(t) - H(t-1)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survDistr$hazard(times = NULL, eps = 1e-06)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{times}}{(\code{numeric})\cr
New time points at which to interpolate.
Values do not need to be sorted or unique, just non-negative.
If \code{NULL}, the object's stored time points are used.}

\item{\code{eps}}{(\code{numeric(1)})\cr
Very small number to substitute zero values in order to prevent errors
in e.g. log(0) and/or division-by-zero calculations.
Default value is 1e-06.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a hazard \code{matrix}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-survDistr-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-survDistr-pdf}{}}}
\subsection{Method \code{pdf()}}{
Computes the probability density function \eqn{f(t)} at the specified time points.
\eqn{f(t)} is the probability of the event occurring at the specific time \eqn{t}.
For constant survival interpolation, \eqn{f(t) = F(t) - F(t-1)}, where
\eqn{F(t)} is the cumulative distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survDistr$pdf(times = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{times}}{(\code{numeric})\cr
New time points at which to interpolate.
Values do not need to be sorted or unique, just non-negative.
If \code{NULL}, the object's stored time points are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a pdf \code{matrix}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-survDistr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-survDistr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survDistr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
